Misc fix
