+++ mikeio/dataset/_dataarray.py
@@ -10,8 +10,8 @@ import numpy as np
 import pandas as pd
 from mikecore.DfsuFile import DfsuFileType  # type: ignore
 
-from ._data_utils import DataUtilsMixin
 from ..eum import EUMType, EUMUnit, ItemInfo
+from ._data_utils import _get_time_idx_list, _n_selected_timesteps
 
 
 from ..spatial import (
@@ -89,7 +89,7 @@ class _DataArraySpectrumToHm0:
         )
 
 
-class DataArray(DataUtilsMixin):
+class DataArray:
     """DataArray with data and metadata for a single item in a dfs file
 
     The DataArray has these main properties:
@@ -1084,20 +1084,32 @@ class DataArray(DataUtilsMixin):
         -------
         DataArray
         """
+        from scipy.interpolate import interp1d  # type: ignore
+
         t_out_index = self._parse_interp_time(self.time, dt)
         t_in = self.time.values.astype(float)
         t_out = t_out_index.values.astype(float)
 
-        data = self._interpolate_time(
-            t_in, t_out, self.to_numpy(), method, extrapolate, fill_value
-        )
+        data = interp1d(
+            t_in,
+            self.to_numpy(),
+            axis=0,
+            kind=method,
+            bounds_error=not extrapolate,
+            fill_value=fill_value,
+        )(t_out)
 
         zn = (
             None
             if self._zn is None
-            else self._interpolate_time(
-                t_in, t_out, self._zn, method, extrapolate, fill_value
-            )
+            else interp1d(
+            t_in,
+            self._zn,
+            axis=0,
+            kind=method,
+            bounds_error=not extrapolate,
+            fill_value=fill_value,
+        )(t_out) 
         )
 
         return DataArray(
@@ -1869,3 +1881,112 @@ class DataArray(DataUtilsMixin):
             return f"values: [{self.values[0]:0.4g}, {self.values[1]:0.4g}, ..., {self.values[-1]:0.4g}]"
         else:
             return ""  # raise NotImplementedError()
+
+
+    @staticmethod
+    def _parse_interp_time(old_time:pd.DatetimeIndex, new_time: Any) -> pd.DatetimeIndex:
+        if isinstance(new_time, pd.DatetimeIndex):
+            t_out_index = new_time
+        elif hasattr(new_time, "time"):
+            t_out_index = pd.DatetimeIndex(new_time.time)
+        else:
+            # offset = pd.tseries.offsets.DateOffset(seconds=new_time) # This seems identical, but doesn't work with slicing
+            offset = pd.Timedelta(seconds=new_time)
+            t_out_index = pd.date_range(
+                start=old_time[0], end=old_time[-1], freq=offset
+            )
+
+        return t_out_index
+    
+
+    @staticmethod
+    def _time_by_agg_axis(
+        time: pd.DatetimeIndex, axis: int | Sequence[int]
+    ) -> pd.DatetimeIndex:
+        """New DatetimeIndex after aggregating over time axis"""
+        if axis == 0 or (isinstance(axis, Sequence) and 0 in axis):
+            time = pd.DatetimeIndex([time[0]])
+
+        return time
+
+    @staticmethod
+    def _get_time_idx_list(time: pd.DatetimeIndex, steps):
+        """Find list of idx in DatetimeIndex"""
+
+        return _get_time_idx_list(time, steps)
+
+    @staticmethod
+    def _n_selected_timesteps(time, k):
+        return _n_selected_timesteps(time, k)
+
+    @staticmethod
+    def _is_boolean_mask(x) -> bool:
+        if hasattr(x, "dtype"):  # isinstance(x, (np.ndarray, DataArray)):
+            return x.dtype == np.dtype("bool")
+        return False
+
+    @staticmethod
+    def _get_by_boolean_mask(data: np.ndarray, mask: np.ndarray) -> np.ndarray:
+        if data.shape != mask.shape:
+            return data[np.broadcast_to(mask, data.shape)]
+        return data[mask]
+
+    @staticmethod
+    def _set_by_boolean_mask(data: np.ndarray, mask: np.ndarray, value) -> None:
+        if data.shape != mask.shape:
+            data[np.broadcast_to(mask, data.shape)] = value
+        else:
+            data[mask] = value
+
+    @staticmethod
+    def _parse_time(time) -> pd.DatetimeIndex:
+        """Allow anything that we can create a DatetimeIndex from"""
+        if time is None:
+            time = [pd.Timestamp(2018, 1, 1)] # TODO is this the correct epoch?
+        if isinstance(time, str) or (not isinstance(time, Iterable)):
+            time = [time]
+
+        if not isinstance(time, pd.DatetimeIndex):
+            index = pd.DatetimeIndex(time)
+        else:
+            index = time
+
+        if not index.is_monotonic_increasing:
+            raise ValueError(
+                "Time must be monotonic increasing (only equal or increasing) instances."
+            )
+        assert isinstance(index, pd.DatetimeIndex)
+        return index
+
+    @staticmethod
+    def _parse_axis(data_shape, dims, axis) -> int | Tuple[int]:
+        # TODO change to return tuple always
+        # axis = 0 if axis == "time" else axis
+        if (axis == "spatial") or (axis == "space"):
+            if len(data_shape) == 1:
+                if dims[0][0] == "t":
+                    raise ValueError(f"space axis cannot be selected from dims {dims}")
+                return 0
+            if "frequency" in dims or "directions" in dims:
+                space_name = "node" if "node" in dims else "element"
+                return dims.index(space_name)
+            else:
+                axis = 1 if (len(data_shape) == 2) else tuple(range(1, len(data_shape)))
+        if axis is None:
+            axis = 0 if (len(data_shape) == 1) else tuple(range(0, len(data_shape)))
+
+        if isinstance(axis, str):
+            axis = "time" if axis == "t" else axis
+            if axis in dims:
+                return dims.index(axis)
+            else:
+                raise ValueError(
+                    f"axis argument '{axis}' not supported! Must be None, int, list of int or 'time' or 'space'"
+                )
+
+        return axis
+
+    @staticmethod
+    def _axis_to_spatial_axis(dims, axis):
+        # subtract 1 if has time axis; assumes axis is integer
+        return axis - int(dims[0] == "time")
